Extract `{prepare,run}_cargo_test` into `test_helpers` module
